# UserProfileService/openapi-profile-service.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UserProfileService",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/profiles/picture": {
      "post": {
        "summary": "UploadProfilePicture",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "profilePicture": {
                    "type": "string",
                    "format": "binary",
                    "description": "The profile picture file to upload"
                  }
                },
                "required": [
                  "profilePicture"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "isSuccess": true,
                  "message": "Operation completed successfully.",
                  "data": "/images/profiles/63764a87-6a94-44be-8e98-9cc9610c8a09.jpg",
                  "errors": [],
                  "statusCode": 200,
                  "timestamp": "2026-07-09T19:30:00.0000000Z"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "isSuccess": false,
                  "message": "Unauthorized.",
                  "data": null,
                  "errors": [
                    "Unauthorized."
                  ],
                  "statusCode": 401,
                  "timestamp": "2026-07-09T19:30:00.0000000Z"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "isSuccess": false,
                  "message": "Resource not found.",
                  "data": null,
                  "errors": [
                    "User profile not found."
                  ],
                  "statusCode": 404,
                  "timestamp": "2026-07-09T19:30:00.0000000Z"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}